moved everything into params and payload files instead of separating them in different files
