--- models/category.go
@@ -1,11 +1,0 @@
-package models
-
-type CategoryParams struct {
-	Query string `url:"query,omitempty"`
-}
-
-func GetCategoryParams(q string) *CategoryParams {
-	return &CategoryParams{
-		Query: q,
-	}
-}
