--- models/category_payload.go
@@ -1,6 +1,0 @@
-package models
-
-type Category []struct {
-	Code  []string `json:"c"`
-	Score float32  `json:"s"`
-}
