--- models/hub_payload.go
@@ -1,7 +1,0 @@
-package models
-
-type Hub struct {
-	Scrubber ScrubberSearchSchemaJson
-	Category Category
-	Berlin   BerlinSearchSchemaJson
-}
