--- models/scrubber.go
@@ -1,19 +1,0 @@
-package models
-
-import "net/url"
-
-type ScrubberParams struct {
-	Query string `url:"q,omitempty"`
-}
-
-func GetScrubberParams(query url.Values) *ScrubberParams {
-	result := ScrubberParams{
-		Query: "",
-	}
-
-	if len(query["q"]) >= 1 {
-		result.Query = query["q"][0]
-	}
-
-	return &result
-}
